+++ R/kinematics.R
@@ -17,3 +17,74 @@ get_velocity <- function(signal) {
 get_acceleration <- function(signal) {
   return(c(NA, NA, diff(signal, differences = 2)))
 }
+
+
+
+#' Get gestural landmarks
+#'
+#' @param signal The displacement signal.
+#' @param time The time of the signal.
+#' @param threshold The velocity threshold (default is \code{0.2}, corresponding to 20 percent velocity.)
+#'
+#' @return A tibble with one row and a column for each gestural landmark.
+#' @export
+get_landmarks <- function(signal, time, threshold = 0.2) {
+  signal_abs_vel <- abs(get_velocity(signal))
+  peaks <- pracma::findpeaks(signal_abs_vel, minpeakheight = 0.1)
+
+  min_1_time <- time[peaks[1,3] - 1]
+  min_1 <- signal[peaks[1,3] - 1]
+  peak_1_time <- time[peaks[1,2]]
+  peak_1 <- signal[peaks[1,2]]
+  max_disp_time <- time[peaks[1,4] - 1]
+  max_disp <- signal[peaks[1,4] - 1]
+  peak_2_time <- time[peaks[2,2]]
+  peak_2 <- signal[peaks[2,2]]
+  min_3_time <- time[peaks[2,4] - 1]
+  min_3 <- signal[peaks[2,4] - 1]
+
+  # Get gesture onset
+  # Formula: threshold * (abs_vel at peak_1 = peaks[1,2] - abs_vel at min_1 = peaks[1,3] - 1)
+  min1_peak1_thresh <- threshold * (signal_abs_vel[peaks[1,2]] - signal_abs_vel[peaks[1,3] - 1])
+  GEST_ons <- approx(
+    signal_abs_vel[time >= min_1_time & time <= peak_1_time],
+    time[time >= min_1_time & time <= peak_1_time],
+    signal_abs_vel[peaks[1,3] - 1] + min1_peak1_thresh
+  )$y
+
+  # Get plateau onset
+  # Formula: 1 - threshold * (abs_vel at max_disp - abs_vel at peak_1)
+  peak1_max_thresh <- (1 - threshold) * (signal_abs_vel[peaks[1,4] - 1] - signal_abs_vel[peaks[1,2]])
+  PLAT_ons <- approx(
+    signal_abs_vel[time >= peak_1_time & time <= max_disp_time],
+    time[time >= peak_1_time & time <= max_disp_time],
+    signal_abs_vel[peaks[1,2]] + peak1_max_thresh
+  )$y
+
+  # Get plateau offset
+  # Formula: 1 - threshold * (abs_vel at max - abs_vel at peak_2)
+  max_peak2_thresh <- threshold * (signal_abs_vel[peaks[1,4] - 1] - signal_abs_vel[peaks[2,2]])
+  PLAT_off <- approx(
+    signal_abs_vel[time >= max_disp_time & time <= peak_2_time],
+    time[time >= max_disp_time & time <= peak_2_time],
+    signal_abs_vel[peaks[1,4] - 1] - max_peak2_thresh
+  )$y
+
+  # Get gesture offset
+  # Formula: threshold * (abs_vel at peak_2 = peaks[2,2] - abs_vel at min_3 = peaks[2,4] - 1)
+  peak2_min3_thresh <- threshold * (signal_abs_vel[peaks[2,2]] - signal_abs_vel[peaks[2,4] - 1])
+  GEST_off <- approx(
+    signal_abs_vel[time >= peak_2_time],
+    time[time >= peak_2_time],
+    signal_abs_vel[peaks[2,4] - 1] +
+      peak2_min3_thresh
+  )$y
+
+  if (nrow(peaks) == 2) {
+      tibble::tibble(
+        min_1, peak_1, max_disp, peak_2, min_3,
+        min_1_time, peak_1_time, max_disp_time, peak_2_time, min_3_time,
+        GEST_ons, PLAT_ons, PLAT_off, GEST_off
+      )
+  }
+}